requirements.txt
boto3==1.16.40
Django==3.1
djangorestframework==3.12.2
pyjwt==2.0.0
requests==2.25.1